requirements.txt
boto3==1.24.96
bcrypt==4.1.3
cryptography==41.0.5
fastapi==0.95.1
httpx==0.25.2
pydantic==1.10.13
PyJWT==2.6.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==1.4.50
sqlmodel==0.0.11
starlette==0.26.1
uvicorn==0.20.0